Revert "Merge remote-tracking branch 'upstream/main' into my-development"
This reverts commit 2ac3744c2ff9013c826fc8f7a0e51e98ef9939e3, reversing
changes made to bafdce7da59b30f7487a63f595a3b56af21eff65.

+++ Intersect.Server/Database/PlayerData/Mute.cs
@@ -70,11 +70,6 @@ namespace Intersect.Server.Database.PlayerData
 
         public string Muter { get; private set; }
 
-        [NotMapped]
-        public bool IsExpired => Expired(this);
-
-        public static bool Expired(Mute mute) => mute.EndTime <= DateTime.UtcNow;
-
         public static bool Add([NotNull] Mute mute, [CanBeNull] PlayerContext playerContext = null)
         {
             lock (DbInterface.GetPlayerContextLock())
@@ -104,8 +99,10 @@ namespace Intersect.Server.Database.PlayerData
             [NotNull] string muter,
             string ip,
             [CanBeNull] PlayerContext playerContext = null
-        ) =>
-            Add(new Mute(userId, ip, reason, duration, muter), playerContext);
+        )
+        {
+            return Add(new Mute(userId, ip, reason, duration, muter), playerContext);
+        }
 
         public static bool Add(
             [NotNull] User user,
@@ -128,10 +125,12 @@ namespace Intersect.Server.Database.PlayerData
             [NotNull] string muter,
             string ip,
             [CanBeNull] PlayerContext playerContext = null
-        ) =>
-            client.User != null && Add(client.User, duration, reason, muter, ip, playerContext);
+        )
+        {
+            return client.User != null && Add(client.User, duration, reason, muter, ip, playerContext);
+        }
 
-        public static bool Remove(Mute mute)
+        public static bool Remove(Guid userId, [CanBeNull] PlayerContext playerContext = null)
         {
             lock (DbInterface.GetPlayerContextLock())
             {
@@ -139,69 +138,24 @@ namespace Intersect.Server.Database.PlayerData
                 if (context == null)
                 {
                     return false;
+                }
+
+                var mute = context.Mutes.FirstOrDefault(p => p.UserId == userId);
+                if (mute == null)
+                {
+                    return true;
                 }
 
                 context.Mutes.Remove(mute);
-
                 DbInterface.SavePlayerDatabaseAsync();
 
                 return true;
             }
         }
 
-        public static bool Remove(string ip, bool expired = true)
+        public static bool Remove([NotNull] User user, [CanBeNull] PlayerContext playerContext = null)
         {
-            lock (DbInterface.GetPlayerContextLock())
-            {
-                var context = DbInterface.GetPlayerContext();
-                if (context == null)
-                {
-                    return false;
-                }
-
-                var mutes = context.Mutes.Where(e => e.Ip == ip && (!expired || Expired(e))).ToList();
-
-                if ((mutes?.Count ?? 0) == 0)
-                {
-                    return true;
-                }
-
-                context.Mutes.RemoveRange(mutes);
-
-                DbInterface.SavePlayerDatabaseAsync();
-
-                return true;
-            }
-        }
-
-        public static bool Remove(Guid userId, bool expired = true)
-        {
-            lock (DbInterface.GetPlayerContextLock())
-            {
-                var context = DbInterface.GetPlayerContext();
-                if (context == null)
-                {
-                    return false;
-                }
-
-                var mutes = context.Mutes.Where(e => e.UserId == userId && (!expired || Expired(e))).ToList();
-
-                if ((mutes?.Count ?? 0) == 0)
-                {
-                    return true;
-                }
-
-                context.Mutes.RemoveRange(mutes);
-
-                DbInterface.SavePlayerDatabaseAsync();
-
-                return true;
-            }
-        }
-
-        public static bool Remove([NotNull] User user)
-        {
-            if (!Remove(user.UserMute))
+            if (!Remove(user.Id, playerContext))
             {
                 return false;
             }
@@ -211,7 +165,10 @@ namespace Intersect.Server.Database.PlayerData
             return true;
         }
 
-        public static bool Remove([NotNull] Client client, [CanBeNull] PlayerContext playerContext = null) => client.User != null && Remove(client.User);
+        public static bool Remove([NotNull] Client client, [CanBeNull] PlayerContext playerContext = null)
+        {
+            return client.User != null && Remove(client.User, playerContext);
+        }
 
         public static string FindMuteReason(
             Guid userId,
@@ -229,14 +186,7 @@ namespace Intersect.Server.Database.PlayerData
 
                 var mute = Find(userId) ?? Find(ip);
 
-                var expired = mute?.IsExpired ?? true;
-
-                if (expired && mute != null)
-                {
-                    Remove(mute);
-                }
-
-                return expired
+                return mute == null
                     ? null
                     : Strings.Account.mutestatus.ToString(mute.StartTime, mute.Muter, mute.EndTime, mute.Reason);
             }
@@ -248,37 +198,22 @@ namespace Intersect.Server.Database.PlayerData
             [CanBeNull] PlayerContext playerContext = null
         )
         {
-            lock (DbInterface.GetPlayerContextLock())
+            if (user.Mute == null)
             {
-                var context = DbInterface.GetPlayerContext();
-                if (context == null)
-                {
-                    return null;
-                }
-
-                if (user.Mute == null)
-                {
-                    user.IpMute = Find(ip);
-                }
-
-                var mute = user.Mute;
-
-                var expired = mute?.IsExpired ?? true;
-
-                if (expired && mute != null)
-                {
-                    Remove(mute);
-                    user.IpMute = null;
-                    user.UserMute = null;
-                }
-
-                return expired
-                    ? null
-                    : Strings.Account.mutestatus.ToString(mute.StartTime, mute.Muter, mute.EndTime, mute.Reason);
+                user.IpMute = Find(ip);
             }
+
+            var mute = user.Mute;
+
+            return mute == null
+                ? null
+                : Strings.Account.mutestatus.ToString(mute.StartTime, mute.Muter, mute.EndTime, mute.Reason);
         }
 
-        public static Mute Find([NotNull] User user) => Find(user.Id);
+        public static Mute Find([NotNull] User user)
+        {
+            return Find(user.Id);
+        }
 
         public static Mute Find(Guid userId)
         {
@@ -301,11 +236,20 @@ namespace Intersect.Server.Database.PlayerData
             }
         }
 
-        public static IEnumerable<Mute> FindAll([NotNull] User user) => ByUser(DbInterface.GetPlayerContext(), user.Id);
+        public static IEnumerable<Mute> FindAll([NotNull] User user)
+        {
+            return ByUser(DbInterface.GetPlayerContext(), user.Id);
+        }
 
-        public static IEnumerable<Mute> FindAll(Guid userId) => ByUser(DbInterface.GetPlayerContext(), userId);
+        public static IEnumerable<Mute> FindAll(Guid userId)
+        {
+            return ByUser(DbInterface.GetPlayerContext(), userId);
+        }
 
-        public static IEnumerable<Mute> FindAll(string ip) => ByIp(DbInterface.GetPlayerContext(), ip);
+        public static IEnumerable<Mute> FindAll(string ip)
+        {
+            return ByIp(DbInterface.GetPlayerContext(), ip);
+        }
 
         #region Compiled Queries
